+++ trixnity-messenger/src/commonTest/kotlin/de/connect2x/trixnity/messenger/viewmodel/roomlist/RoomListViewModelMultiAccountTest.kt
@@ -854,7 +854,7 @@ class RoomListViewModelMultiAccountTest {
             selectedRoomId = MutableStateFlow(RoomId("!roomId")),
             onRoomSelected = onRoomSelectedMock,
             onCreateNewRoom = mock(),
-            onUserSettingsSelected = mock(),
+            onShowAccounts = mock(),
             onUserProfileSelected = mock(),
             onOpenAppInfo = mock(),
             onSendLogs = mock(),